Extract visualization logic
**  No behavioral change intended

Issues: [#7]
